requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
matplotlib==3.9.0
kafka-python==2.0.2
SQLAlchemy==2.0.31
python-dotenv==1.0.1
PyYAML==6.0.2
joblib==1.4.2
sqlalchemy>=2.0
psycopg2-binary==2.9.9
